requirements.txt
python-telegram-bot
aiohttp
azure-identity
azure-keyvault-secrets
python-dotenv